Valearnis/requirements.txt
asgiref==3.5.2
Django==4.1.4
django-environ==0.9.0
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
future==0.18.2
pytz==2022.6
serializers==0.2.4
sqlparse==0.4.3
